RunnableListenerHelper: Performance improvements when run once
Rather than have an AtomicBoolean which requires a native context switch to do the `compareAndSet`, we now have a volatile boolean that we set while synchronized.  This is a performance gain because we will have to synchronize in doCallListeners now anyways (which we now synchronize in earlier).
This is a minor performance improvement, but since it impacts all `ListenableFuture` implementations it's a note worthy one.