5.5 Patch for scheduler shutdown low priority task completion
This fixes an issue in PriorityScheduler and SingleThreadScheduler where invoking .shutdown() will not wait for low priority tasks to complete before the pool shuts down (only waiting for High priority tasks).
Starvable tasks are considered starvable and thus there is not consideration in waiting for them to complete before shutdown occurs.