ListenableFuture.map / flatMap improved interrupt ability
Previous to this we would only be able to interrupt / cancel the mapped future if the source future has not begun.
With this change if an interrupt is requested we will interrupt / stop the mapping process.

The change to lookup / set the current thread before the function is processed adds about 5% to 10% additional overhead in mapping futures.
Even without setting the running thread the cancel logic is improved to at least stop chains of mapped futures from continuing.  That said, I believe right now the overhead is worth the more likely to be expected behavior of sending a thread interrupt to the processing mapping function.