ExecutorLimiter avoid limiting listeners / monads associated to futures
With threadly 5.0 we made ListenableFuture's even more useful with monads map and flatMap functions.
However in doing that, it becomes easier to have issues where you are piggy backing the thread that did the original execution without thinking about it.
Normally this is no big deal if the pool you are on is big enough, or the operation you are doing is fast enough.  But if you are on a constrained pool, or the listener may be slow due to DB actions / etc, it could end up fully consuming a pool you did not expect.

This is designed to help address this concern by encouraging the use of limiters rather than entire pools to limit access to resources.
By allowing limiters to not include those operations you can more easily use a limiter to garud such resources, but have it backed by a larger pool to handle these slower / more general tasks.