Added `ThreadRenamingPriorityScheduler`
This is similar to the other variations that already exist, but for the `PrioritySchedulerService`.

The only delay was because `AbstractPriorityScheduler` is not really useful for an implementation like this.  We potentially could have minorly made this better by using default functions in the interface, but I decided to keep this simple for now.