Clock: Use central pool rather than create another thread
This thread is blocking almost always, and we can do the same action on the central pool.
There is a slight behavior modification in that stopping the Clock wont ensure the update thread / task has completed.  But I think that was a detail that really was only important for our unit tests.