CentralThreadlyPool: Added "isolatedTaskPool"
This implementation is to avoid the discreet comment in Clock (which was reverted anyways) around the memory overhead of the single thread pool.
This hopefully puts the details into the API in a way that will allow users to pick the pool that makes the most sense for them (even though both technically work).