Common interface `RunnableRunnableContainer`
Typically something that contains a runnable, itself is executable.  This interface combines that concept, and will facilitate in making some future changes go easier / cleaner.
The only downsides to this are the javadocs pollution, as well as a VERY VERY MINOR potential for performance reductions by further polluting the Runnable implmenetation instances (of specifically TaskWrappers).