Adds an additional option to optimize the listener executor on a per-listener basis
This adds four new functions to `ListenableFuture` where the executor for the listener may be optimized away.
This can allow the listener to avoid the need to re-queue in the executor and get thread time _again_ before it can be executed.
However, if the listener was complex, or wanted to be run concurrently this will cause issues, and thus why we can't default enable this type of operation.