ListenerHelper / RunnableListenerHelper: execute async listeners before in-thread
The thought behind this is listeners that have an executor to execute on don't care about the order they were added to the listener.  Because in-thread listeners may delay their execution on to the pool, we keep them seperate and execute them first.
As written this increases the code duplication a bit more than I wish, particularly considering how popular this code is.  But it does improve the behavior as described, and does so without the impact of a Map.