ListenableFutureTask & SettableListenableFuture improvements to removing executor reference for GC
I was double checking if there was a possible race condition where we might be checking the executor reference and the done state.
That seems correct, particularly since we are checking done first.
But in `ListenableFutureTask` we can remove the reference for GC earlier.
and in `SettableListenableFuture` we were forgetting to clear the reference at all.