ListenableFuture: add `throwMap` capabilities
In 5.10 we changed it so that exceptions thrown from the mapper function get reported as "uncaught" exceptions to `ExceptionHandlers.handleException(Throwable)`.
This makes sense for most cases, but it was discovered that some cases actually do want to throw or translate into an exception.  So this adds `throwMap` to provide the pre-5.10 behavior of `map`.