5.12 Release to fix recent regression when using ListenableFuture optimizations
A highly unlikely race condition was introduced where a future may complete while the optimization is being checked resulting in an on-thread execution that was not desired.
If `SingleThreadIfExecutorMatch` was passed and the entire optimization check statement has completed before the thread is suspended, and then only resumed once the future has fully completed, the provided listener will invoke on the calling thread even though not desired to do so.
This is fixed by extending this optimization capabilities to the `RunnableListenableHelper` where this decision is automaically made.