Internal listener trick to get early execution
Recently we optimized listener execution to invoke listeners with Executors first.  This was to allow these async operations to not be blocked by in-thread listeners.
In this commit we game the new system by providing a `SameThreadSubmitterExecutor` so we can cut in line, but still execute in-thread.
This allows our library behavior to be more consistent, reducing the impact with potential user added listeners.
In all cases this was done, there should be a benefit, as well as no risk to previously expected behavior by the listener operation being very fast.