Added `FutureUtils.executeWhile`
This is similar to `scheduleWhile` except it does not need a scheduler.  In fact it does not even take an executor (instead using `ListenableFuture`'s to leapfrog to the next).
`scheduleWhileTaskResultNull` was deprecated because it seemed like it did not add much value (would not be hard to replicate the functionality, not worth cluttering up the API, particularly since it would need to be added here too).