RunnableListenerHelper: Inline `runListener` function
I actually think that the code is cleaner having this as a function, even though all three calling places have slight variations of the behavior.
The motivation for this change comes from the increased use of `map` and `flatMap` to transform ListenableFuture's.
If futures are mutating things with already complete results the stack can become quite large, so anything we can do to reduce the threadly noise is good.