ListenableFuture: Simplify flatMap implementation
We don't need this additional inner class, we can provide the resulting SLF directly as a callback.
The only concern in doing this was potential cancel states not being communicated.  Because of that the instanceof check was moved to SettableListenableFuture's handleFailure.  I think this behavior is likely what most people would expect anyways (even though I hate the `instanceof` check).