Add `ListenableFuture.mapFailure` / `flatMapFailure` (#232)
These are similar to `throwMap` and `flatMap` in that they allow you to transform an async result from one future into another.
However instead of mapping one result into another result, these map the failure conditions into another failure condition, or back into the expected result type.
This means they wont be invoked if the future is not in an error condition (similar to how `map` wont be invoked if the future IS in a failure condition).  That is why they can only map the failure into either another failure, or back into the expected result type.