WatchdogCache: reduce memory overhead when used with a wide variety of timeouts
Because our timeout field accepts a millisecond resolution, we may end up with a lot of garbage / overhead if people start invoking in with actual millisecond level resolution.
This commit makes the class safer by defaulting everything to be resolved within 200 milliseconds, but can be constructed with other values to allow different levels of granularity.