Minor Watchdog/WatchdogCache improvements
Micro improvements after looking through the code again.
An additional active timeout for WatchdogCache was considered, but unclear if there is a benefit since the schedule nature of the watchdog checker kind of already providers a timeout like capability.  There is still always the possibility that a cache will be checked at the wrong point and be removed even though it's normally active, but this minor churn is probably not significant enough to warrant the extra code.