RunnableListenerHelper: Micro optimization for case where single listener is specified
After benchmarking this shows to be a bit of an improvement when a single listener is used, with no detectable loss in performance when 2 or 4 listeners are used.
In theory this adds an extra copy in cases where 2 listeners are used, but this change allows us to avoid a copy up to 4 listeners.  With 4 listeners the performance is actually better.
The extremly minimal loss at 2 listener cases is expected to be overset by the gains for the very common single listener case (ie when mapping futures).