SettableListenableFuture: Fix for not completing into cancel state when cancel is overriden
In FutureUtils CancelDelegateSettableListenableFuture we may not cancel `this` future if the delegate future can't be canceled.
This behavior may make sense there, but we also need to ensure that the failure condition the SettableListenableFuture witnessed is not lost.  So if that cancel fails, we will try again with a function that can't be overriden to ensure the future absolutely completes.