Add `StackTracker`
This resolves #187 by providing the ability to indicate when a stack should be recorded.  It's not exactly the same in that you can't provide the arbitrary stack traces to be de-duplicated (though the internal `ComprableTrace` could easily provide that).  Instead the API is simplifed into just calling a function when you want the stack to be recorded, and then allowing simple export / reset functionality.