SettableListenableFuture: FutureCallback performance improvements
This provides a lambda implementation for FutureCallbacks that consider the internal set state.
For normal results this seems to be basically performance neutral (maybe a slight edge, more edge possible if we could avoid synchronization).
For error results however this is significantly better.  Not needing to wrap error conditions in an `ExecutionException` result in ~ a 10x improvement on my system.