SettableListenableFuture: Don't create object for result lock
Internally we know that `RunnableListenerHelper` wont synchronize on `this` (it has very specific lock needs so likely never will).  So let's reference it as our result lock and avoid constructing an object just to lock (while still avoiding also needing to synchronize on `this` which I feel may be dangerous for a library to do).