Profiler: Fix bug when specifying `SameThreadSubmitterExecutor` for profile executor
If this executor was provided previously the profiler would collect samples, but never be able to stop.  Instead it would hold on to the lock causing stop calls to just block.
The handles the `SameThreadSubmitterExecutor` logic uniquely (for reasons described in a long comment).  That does mean that if in-thread execution occurs from a different implementation this will still be broken (yuck), but I am not seeing a way to deal with that right now.