To also be included in 5.24, SettableListenableFuture callback stack reductions
The previous commit unconditionally called into `addListener` thus increasing the stack by one.
This inlines addListener to reduce that stack frame, but also adds in an optimistic check for a set result / listener.  This adds very minimal overhead for `null` result cases, but gives us to opportunity to not only make it potentially faster for the other cases, but also reduce their stack depths as well.