Profiler: Treat thread name changes as different threads
Previously if a thread changed its name, we would profile it under the same thread based off the id.  This considers now id and thread name, if the name changes it will be the same as if it was a different thread.
This is valuable for things like the `CentralThreadlyPool` where a single thread may only indicate it's true use by its name.  This was even worse in that previously the profile would have shown only the first name it witnessed as.

In addition to this behavior improvement, there is some minor code optimization and cleanup included.