Profiler: More changes towards thread names
I worried that by treating thread names as different threads it might give an unrealistic expectation of the atomic operation of checking the stack and the name.
Since we can't automically get both pieces of information it's possible for the name to change and be represented with a stack that did not occur while having that name.

This changes puts us back to tracking by the thread identifier.  However the logic is improved compared to before in that we will now sample the same every time.  If we see multiple names for a given thread they will all be reported.
In addition the `dump` operation is slightly more expensive, because it will now sort the threads based off their name (hopefully putting similar threads close together)