Profiler: Improvements to reporting idle threads
This requires a bit of code, but should be very low impact / fast to make use of.
Most the code comes in a static block of `Profiler` where we need to start multiple pools and record their thread states.
We then use these recorded stack traces to know what idle threads of these implementations look like.
This is particularly useful for threadly pools, but currently also supports ThreadPoolExecutor idle threads (maybe more could be added in the future).
The primary goal was to have these stack traces reduced into very easy to read descriptions.  In addition this provides a minor memory usage improvement as these traces are witnessed.