TestCondition: Allow generic type T supplier / predicate
This improves `TestCondition` with a third way it can be used.  Allowing you to specify a supplier of type `T` and a predicate that test the results, we can report the final (failing) value.  This can allow for better debugging as you can see the final state (beyond just a boolean if it was passing or not).
Unit tests use of `TestCondition` was audited, trying to switch to one of the prefered methods of use now (to allow better debugging hopefully).
I would like to get rid of the `TestCondition()` constructor, and the overriding of `get()` but the fact that you can't reference `this` in calling super constructors does provide some cases where it's hard to get away from allowing it.