One limiter task for all keys
This converts the design of having one limiter task per key, to one task for all keys.
This reduces memory overhead, as well as likely computation overhead as lots of tasks are now managed by just one.
The risk would be if there is so many keys that the limiter checker can't keep up.  That may occur at 10's or 100's of millions of keys.  But I suspect the old design likely would not have done well at that scale anyways (due to the larger memory overhead).