KeyDistributedExecutor: Switch off StripedLock
This removes the use of StripedLock and instead relies on `ConcurrentHashMap.compute` for guarding a given key from concurrent operations.
This makes the code simpler in some ways, removing constructors (that were likely confusing), and using less internal tooling.  But does result in the annoying array's to capture results of operations we needed in the `compute` block.