ListenableFuture.flatMap cancel async returned future fix
This improves the behavior when ListenableFuture.flatMap is used, and the mapped future is canceled AFTER the source future has completed, but before the returned future from the mapper has completed.
Previously the returned future from the mapper would not be canceled, though if the source future was already complete before the mapping was attempted this would be canceled.
This change makes the behavior more consistent and ensures that futures intended to be canceled are in fact canceled.