ListenableFuture: add getRunningStackTrace()
This commit adds the ability to get the stack trace for the thread currently processing the future (or a future which we need the result for currently).
This thread may be `null` if not started, or already completed, or because of a potential inconsistent state as futures are processed.  Despite the lack of guarantees, this should provide some basic abilities for debugging.

The big unfortunate fact here is the inclusion of Unsafe.  We have resisted it for so long, but this brings it in for very frustrating reasons.
We need to be able to access the `runner` field in java.util.concurrent.FutureTask, otherwise we need another native invocation to set the Thread reference for this implementation.
In theory this should be a simple:
  Field field = FutureTask.class.getDeclaredField("runner");
  field.setAccessible(true);
However java9+ restricts reflection in a way that this wont work.  I tried building jigsaw modules to `export java.base` in a way to allow this, but could never get it to work.

Instead this commit adds in `UnsafeAccess`, which gets a handle on Unsafe, and then uses it to make AccessibleObject.setAccessible0 visible.  `setAccessible0` allows us to modify access while bypassing the SecurityManager permission checks.  Believe it or not, but the requirements to get Unsafe working in java8 and java9+ seemed to be way simpler than getting SecurityManager to allow us to otherwise modify the permissions.  Though the resulting code code is much more complicated.