SettableListenableFuture: remove finishCompletion function
Having this function to be overloaded is the cleanest code structure wise.
However this stack path is used for every ListenableFuture.map, making a point where we are trying to keep the stack short a little longer.
So instead this duplicates the code, and has some minor visibility changes so that we can avoid this function and instead just duplicate these two lines.