FutureUtils.executeWhile improvements to tracking the running thread
This moves where the running thread is tracked to inside the callback loop.  This allows the running thread to be tracked while the Predicate is checked.
In addition this includes some code simplifications for the executeWhile case that were not as possible for the scheduleWhile.  I removed the TODO to de-duplicate because I am now convinced the problems are different enough to not warrant merging.