FutureUtils.makeResultListFuture / makeCompleteListFuture deterministic order
This improves the previous implementation which provided a resulting list in a non-deterministic order into one which returns the results in an order matching the original Iteration order of futures.
Micro-benchmarking shows this to actually be a faster implementation.
Building a success / failure result list is still in a non-deterministic order, but does have the same performance improvements.