FutureUtils.makeSuccess/FailureList order improvements
This is similar to the last commit in that it changes the order of the results to be determinstic based off the input order.
Unlike the previous commit this does incur a minor performance penalty to do (particularly for the `failure` case due when few failures are reported).
However this has been micro-benchmarked heavily and optimized to where the order could be worth it.
What I like most about this is the removal of the confusing and awkward `count` variable which may be updated from the super and thus may be fragile.