5.32 release: mapped CancellationException message fix
In 5.31 we added the ability to provide debugging information in the CancellationException of a SettableListenableFuture
However this information could be easily lost through map / flatMap / failureMap / etc operations.
This fixes that, trying to make this logic a bit more to the core of how ListenableFuture operates.

I don't particularly like how this state must be communicated, and the additional checks and complication this adds, but I don't have a better idea of how to communicate this debugging information / state.