PriorityScheduler.workerIdle interrupted check optimization removal
After benchmarking this further I don't think the benefit is as high as I want considering the potential risk.
Risk is low, but there is a risk that if an idle worker thread is interrupted multiple times this could start to tight loop polling for tasks (as park would immediately unblock).
This changes the interrupted check to be unconditional at the end, and before we park, making sure this is never possible.  It has the advantage of tracking less state, but may require additional native calls than it did previously.  Still, overall, a little safe, and not really much slower.
Most of the performance loss is in the unconditional check in the `finally` block, where we otherwise might have avoided this.