ExceptionUtils improvements for getRootCause/hasCauseOfType(s)
This provides cicular cause detection for the root Throwable provided into each function.  This otherwise would have resulted in an infinite loop.
We are still only providing detection to the root Throwable, so other circular chains can exist.  These will still result in infinite loop, but the detection of such may be more overhead than we want to accept.