ListenableFuture.addCallback / addListener replaced with `callback` and `listener`
The motivation around this change is not really to make the function names more concise.  Instead it's more to allow these functions to return `this`.
By returning `this` for `listener` and `callback` we can continue to use our more functional style similar to the other `map` and `flatMap` like functions.