Add ListenableFuture.resultCallback/.failureCallback
This provides an easier way to add a callback when you just care about either the result or failure stats only.
In both of these they accept a `java.util.function.Consumer` of the appropriate type.

Because this makes it easier to handle these cases, classes which used to assist with this are now depreciated:
* AbstractFutureCallbackResultHandler (Now implement a `Consumer` directly)
* FutureCallbackResultHandler (Consumer can now be provided to `resultCallback` directly)
* AbstractFutureCallbackFailureHandler (Now implement a `Consumer` directly)
* FutureCallbackFailureHandler (Consumer can now be provided to `failureCallback` directly)
* FutureCallbackExceptionHandlerAdapter (`ExceptionHandler` can be provided as consumer directly to `failureHandler`)