Add FilteredStackProfiler.
This provides a fairly simple way to profile a specific part of a busy
application.

I used java.util.regex for matching stack trace elements, which is
usually something I'd avoid. However, it didn't seem like such a bad
option here since the strings to be matched will always be small and
people are unlikely to try to use patterns with pathological performance
characteristics.

Another option might be to take a `Predicate<StackTraceElement>`, which
is more flexible but introduces the liability of calling back into user
code.