ExceptionUtils improved safety in cycles with exception causes
This adds complete cause cycle detection for:
* getRootCause
* getCauseOfTypes
* hasCauseOfTypes
* getCauseOfType
* hasCauseOfType

This implementation will try to optimize for no cycle in the chain, trying to defer to the old logic until the cause chain gets too deep.
Once the cause chain is deep the process is typically restarted, storing the causes into a structure so they can be checked for a possible cycle.

If a cycle exists or not, regardless of depth, the behavior should now be more consistent (vs just looking for cycles to the start instance).