Add FlowControlledProcessor
This is a type of limiter, but not in the form of a pool wrapper.  Instead this abstract class will pull for tasks when it is ready to accept more.
This is to make it easy to do processing which loads data (for example reading from a file).

The class will invoke `hasNext()` to check if the class has more work to execute.  Assuming it does, and assuming there is capacity to execute it, the class will then invoke `next` to start execution.
As tasks complete the relevant `handleResult` or `handleFailure` will be invoked, providing the results of the execution (in order if desired).

Possibly the most awkward aspect of this implementation is when things will be called concurrently vs not.  `next()` wont be called concurrently due to the need to check for execution before invoking.  But `handleResult` / `handleFailure` may be called concurrently (unless in-order processing is specified).