Move BlockingQueueConsumer to new processing package
This deprecates org.threadly.concurrent.BlockingQueueConsumer.  I considered strongly providing no replacement for this class.
It was obviously missed in the 5.X API changes, still having an unecessary interface.  It also had a more primitive form of error handling.  Poor javadocs.  And generally showed a sign of a forgotten class from an earlier form of this project.

With the recent addition of `Poller.consumeQueue` which can take any `Queue` type and has a much cleaner API, I wondered if this was worth maintaining.  The big difference between these is how the empty state is handled, and why I ultimately decided to provide an updated version.
In Poller after an empty state it will wait to poll for more items.  This could be high overhead on active, but still easily drained, queues.  Since there still seems a case with BlockingQueues where some error handling, and thread state handling, are a nice benefit, it seemed worth keeping the minimal class.