Added FutureUtils.makeFirstResultFuture
This is similar to ScheduledExecutorService.invokeAny, but has not historically been something threadly has provided.
It was only recently I was convinced of a use case where we would want to take the first result.
Rather than make it part of our pools like java.util.concurrent does, I think it makes more sense to put it in FutureUtils, so we can be flexible to collections of futures regardless of their sources.