ExceptionUtils: Try not to swallow StackOverflowError
This resolves #237 by setting StackOverflowError's as a volatile, then having a regular task check for them to be set.
It's important that handleException never throws an exception, however due to StackOverflow's making it hard to even log, that may result in the exception / problem being swallowed.
Setting a volatile does not expand the stack, so we can pull to see if any occured.  This may result in some exceptions being lost, however is better than getting no indication at all.