Revert "Add `ThreadlyCompletionService`"
This reverts commit c8eda2b338d731949140a798e549e1f4b22f12e1.

After thinking about this more, this probably does not make sense as an API expansion.
Rather than polling for completed futures Threadly is better designed to add a listener to the future which can invoke a function (providing the result even).
The only argument I could see that a poll design is more ideal is that you get the queuing behavior rather than a forced push on completion.  However this behavior too could easily be replicated easily.  A queue may be trivial to add, but perhaps even simpler you can provide a limited thread pool which is also in-effect a queue.