AbstractPriorityScheduler QueueSet.getNextTask optimization
Optimization in getNextTask to avoid clock check for checking low priority tasks.
Previously we checked nextHighTask.getScheduleDelay() first, which will update the Clock if not ready, we now do this check last in the OR sequence.
The first check for the low priority task is now done relative to the high priority task time instead of the Clock for performance.
The comment has been updated to describe the needed constraints for this to work.