PriorityScheduler: Reduce Clock updates for Low / Starvable tasks
This will still update the clock when the task is submitted to ensure a task never runs early.
However for Low and Starvable tasks this will avoid updating the clock for execution.

This does mean it comes with the risk that Low / Starvable priority tasks wont ever execute if:
   * The user disables the Clock update thread manually
   * No other Clock updates are happening (ie no additional scheduled tasks, no other library or user actions needing to update the clock)

So this is not exactly risk free, but it can significantly improve performance impacts from Low / Starvable priority tasks.  Not just when they are executed, but if they are next for execution and we need to see how long to park the thread.

I have done extensive JMH benchmarking to make sure there is no negative performance impacts from this.  Except for code complexity, there should be no negative trade offs I can find.

I did not want to store the state as a boolean in the TaskWrapper because:
   * It would increase the per-task memory overhead
   * It increases the state checks needed (was being seen as a performance impact in JMH)

Instead we use different class types to indicate this behavior state, which does require a fair bit more code.  I also wanted to prevent additional stat checks when submitting tasks, so I had to expand that code too.