RateLimiterExecutor: Use non-locking implementation
This removes the `permitLock` synchronization and instead does `compareAndSwap` to a AtomicLong (storing a double).
This may overall increase CPU overhead as the operation is retried, but micro-benchmarking shows a notable performance improvement when there is thread contention.
When there is no thread contention the overhead of the CAS / Double translation seems to be very low.  Benchmarks at worse show the overhead being the same, and in most cases show some minor improvements.