Report mapper in exception message when flatMap mapper returns null
In general these mappers should only return a `ListenableFuture` instance.  If they are intending to return a null result it's better to use `map`, or if needing flatMap `FutureUtils.immediateResultFuture(null)`.
Previously this would throw a fairly confusing NullPointerException, maybe even expecting a bug in the library.  With this change we log the `mapper.toString()` in the exception message to help find the bug in the users codebase.