PriorityScheduler: Minor improvements to Worker loop
This improves workerIdle and the queue update notification by avoiding to need to check the `shutdownFinishing` state.
Instead this state is communicated by setting the max pool size to zero on shutdown.

In addition the Worker no longer extends AbstractService.  Despite the minor code increase, it's worth it due to the qty of workers which might exist.