Profiler: Control dump production when start with a timeout is used
When start with a timeout is used a ListenableFuture is returned.  This future completes with the the dump() results at the time of completion.
If the dump is not needed or not desired (potentially from large profiles), this can now be controlled with these additional constructors.
Providing `Profile::dump` will replicate the current behavior, while alternatives like `(p) -> null` or other dump configurations may be useful.