OrderedExecutorLimiter: Add the ability to adjust the limits at runtime
This functionality from the defered to ExecutorLimiter needed to be manually extended.
Also added was a protected `checkTaskCapacity()` function to allow a good place to override if the capacity is to be adjusted dynamically.

In ExecutorLimiter canSubmitTaskToPool() was renamed to taskCapacity().  This seems to be a better name and is a little less confusing compared to `canRunTask()`.

Javadocs were improved and unit tests were extended to cover the new functionality.