OrderedExecutorLimiter: Fix NullPointerException if ListenableFuture is canceled
The task used for sorting the queue is cleared after the future completes.
In order to fix this OrderedExecutorLimiter provides a ListenableFutureTask that wont clear the Callable task until we ensure the future has been removed from the queue (and thus wont be compared against again).  PriorityBlockingQueue is locked and can provide this ensurance.