ConfigurableThreadFactory: Add ability to set a consumer to be notified on thread creation
This adds to ConfigurableThreadFactory and the extending class ThreadReferencingThreadFactory the ability to be notified of threads that are created.
This can be useful for debugging or other monitoring code which needs to know about threads related to specific pools.
The javadocs of ControlledThreadProfiler were updated as this is a perfect example of how this can be useful.